cleanup stage 1 - redistribute the set up qkpt and upload coulomb V tasks in minimal number of functions
