+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(gpu-lib CXX)
   
 add_library(accel-lib cuda_common.cu cuda_check.cpp cublas_routines_prec.cu
-        cu_compute_Pq.cu cugw_qpt.cu cu_routines.cu cu_symmetry.cu )
+        cu_compute_Pq.cu cugw_qpt.cu cugw_qkpt.cu cuhf_utils.cu cugw_utils.cu cu_symmetry.cu )
 set_property(TARGET accel-lib PROPERTY CUDA_ARCHITECTURES ${GPU_ARCHS}) #also check the same parameters in solvers
 target_include_directories(accel-lib PUBLIC .)
 target_link_libraries(accel-lib CUDA::cudart CUDA::cuda_driver CUDA::cublas CUDA::cusolver)
